wsinterfaceproject/requirements.txt
azure
azure-storage-queue
Pillow
NumPy
django==2.1.7
opencv-python
